+++ CHANGELOG.md
@@ -7,6 +7,7 @@ unreleased
 - `dropinInstance.teardown` will return a promise if no callback is provided
 - `dropin.create` will return a promise if no callback is provided
 - Fix error thrown in console when removing fields with card overrides
+- Add `paymentOptionSelected` event
 
 1.3.1
 -----